Restructure: Move more platform specific file to platform-specific folder. Create a different directory for factory files.
