+++ src/drivers/factory/CMakeLists.txt
@@ -1,15 +1,14 @@
 # List of source files
-set(INTERFACES_SOURCES
-    IPin.cpp
-    PinBase.cpp
+set(FACTORY_SOURCES
+    PinFactory.cpp
 )
 
 # Add these files to a library
-add_library(interfaces ${INTERFACES_SOURCES})
+add_library(factory ${FACTORY_SOURCES})
 
 # Other libraries needed
-target_link_libraries(interfaces 
-    PRIVATE 
+target_link_libraries(factory 
+    PRIVATE
         common
         io
         printf
@@ -19,7 +18,7 @@ target_link_libraries(interfaces
 )
 
 # Include directories
-target_include_directories(interfaces
+target_include_directories(factory
     PUBLIC                        # make these include paths available others that link to this library
         ${CMAKE_SOURCE_DIR}/src
 )
