+++ src/drivers/motion/DriveFactory.cpp
@@ -2,7 +2,7 @@
 
 #include "common/assertHandler.hpp"
 #include "common/Trace.hpp"
-#include "drivers/interfaces/PinFactory.hpp"
+#include "drivers/factory/PinFactory.hpp"
 #include "drivers/motion/DriveFactory.hpp"
 #include "drivers/stm32f3discovery/timers/GeneralPurposeTimer.hpp"
