+++ src/drivers/stm32f3discovery/common/CMakeLists.txt
@@ -1,5 +1,6 @@
 # List of source files
 set(DRIVERS_COMMON_SOURCES
+    IrqHandlers.cpp
     Stm32f3CriticalSectionGuard.cpp
 )
