+++ src/drivers/stm32f3discovery/common/registerArrays.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "defines.hpp"
+#include "common/defines.hpp"
 #include "stm32f303xc.h"
 
 namespace IO
