+++ src/drivers/stm32f3discovery/io/ExtiPin.cpp
@@ -1,6 +1,6 @@
 #include "common/assertHandler.hpp"
-#include "common/registerArrays.hpp"
 #include "drivers/interfaces/pinBank.hpp"
+#include "drivers/stm32f3discovery/common/registerArrays.hpp"
 #include "ExtiPin.hpp"
 
 std::shared_ptr<ExtiPin> ExtiPin::Create(const ExtiPinInitStruct &pin_init_struct)
