+++ src/drivers/stm32f3discovery/leds/leds.cpp
@@ -1,5 +1,5 @@
+#include "drivers/factory/PinFactory.hpp"
 #include "drivers/interfaces/PinBase.hpp"
-#include "drivers/interfaces/PinFactory.hpp"
 #include "leds.hpp"
