+++ tests/blinky.cpp
@@ -1,8 +1,8 @@
 #include "common/assertHandler.hpp"
 #include "common/delay.hpp"
+#include "drivers/factory/PinFactory.hpp"
 #include "drivers/interfaces/IDigitalOutputPin.hpp"
 #include "drivers/interfaces/PinBase.hpp"
-#include "drivers/interfaces/PinFactory.hpp"
 #include "drivers/stm32f3discovery/io/GpioPin.hpp"
 #include "drivers/stm32f3discovery/mcu/mcuInit.hpp"
