+++ tests/pwm.cpp
@@ -4,8 +4,8 @@
 #include "common/Trace.hpp"
 #include "common/assertHandler.hpp"
 #include "common/delay.hpp"
+#include "drivers/factory/PinFactory.hpp"
 #include "drivers/interfaces/ITimer.hpp"
-#include "drivers/interfaces/PinFactory.hpp"
 #include "drivers/interfaces/pinBank.hpp"
 #include "drivers/io/GpioPin.hpp"
 #include "drivers/leds/leds.hpp"
