+++ tests/usart.cpp
@@ -1,9 +1,7 @@
 #include "common/Trace.hpp"
 #include "common/assertHandler.hpp"
 #include "common/delay.hpp"
-#include "drivers/interfaces/PinFactory.hpp"
-// #include "drivers/io/GpioPin.hpp"
-// #include "drivers/usart/UsartPin.hpp"
+#include "drivers/factory/PinFactory.hpp"
 #include "mcuInit.hpp"
 #include "printf.h"
 
@@ -12,14 +10,6 @@ int main()
     SystemInit();
 
     ASSERT(isSystemInitialized);
-
-    // UsartPinInitStruct pinInit = {};
-    // pinInit.pin_name           = IO::ePin::IO_UART4_TX_PRINT;
-    // pinInit.alternate_function = IO::eAlternateFunction::IO_AF5;
-    // pinInit.baud_rate          = USART::eBaudRate::USART_BAUD_RATE_115200;
-
-    // [[maybe_unused]] PinBase const *usart_print_pin = PinFactory::CreatePin(
-    //     IO::ePinType::IO_PIN_TYPE_PRINTING_USART, pinInit);
 
     int i = 0;
